added a new class PhoneComLogger to set console and file loging
